Assistant checkpoint: Add social footer and align CTA buttons
Assistant generated file changes:
- templates/home.html: Add social media footer and align CTA buttons, Update other feature boxes, Update Fun Assistant box, Add social media footer
- static/css/custom.css: Add social media styles

---

User prompt:

Ok, I am seeing nothing social on this home page.  Should we be linking to all our social accounts (that are relevant in the event world?)?  At least get the buttons and functionality there and I will add the custom links.  In terms of the CTA buttons in the boxes that you added, make sure they are aligned.