Assistant checkpoint: Update spacing and add 'Add Event' to navigation
Assistant generated file changes:
- static/css/custom.css: Add spacing between search and links
- templates/partials/navbar.html: Add 'Add Event' to navigation

---

User prompt:

Ok, there is lots of padding below the image and above the "Find the Best Events Near You - Rated for Fun!" text.  Can you cut that in half?  Also, everyting below that seems a little scrunched....meaning if I was using my fingers to click into the "Enter your location" field I could accidentally hit the "View Full Map" link...don't over do it and just, maybe put about 50% more space between those two elements.  Also, we need to add "Add Event" to the hamburger menu on mobile and in the main nav bar on desktop to the left of "Sign In".  If you click on "Add Event" and you are not signed in, you are redirected to login page and